Refactor workspace edit line module
